Nearly completed the Call Caonverter for FCs. A few small fixes need to be done, but I will do.
I'm also thinking of modifieing the S7FunctionBlockRow, so that the Call Parameters are stored in another way,
so it's easier to handle them.

+++ LibNoDaveConnectionLibrary/PLCs/S7_xxx/MC7/Helper.cs
@@ -888,7 +888,7 @@ namespace DotNetSiemensPLCToolBoxLibrary.PLCs.S7_xxx.MC7
                         return Result;
                         break;
                 }
-            Result = Result + Convert.ToString(tmp / 8) + "." + Convert.ToString(tmp % 8);
+            Result = Result + Convert.ToString(((int)tmp / 8)) + "." + Convert.ToString(tmp % 8);
 
             return Result;
         }
